first draft of checkbox that auto submits
